ref: move all entities to new folder resources
