+++ data-source.ts
@@ -1,12 +1,13 @@
 import 'reflect-metadata';
 import { DataSource } from 'typeorm';
-import { Album } from './src/album/entities/album.entity';
-import { Artist } from './src/artist/entities/artist.entity';
-import { User } from './src/user/entities/user.entity';
-import { Favorite } from './src/favorites/entities/favorite.entity';
-import { Track } from './src/track/entities/track.entity';
+import { Album } from './src/resources/album/entities/album.entity';
+import { Artist } from './src/resources/artist/entities/artist.entity';
+import { User } from './src/resources/user/entities/user.entity';
+import { Favorite } from './src/resources/favorites/entities/favorite.entity';
+import { Track } from './src/resources/track/entities/track.entity';
 import * as dotenv from 'dotenv';
 dotenv.config();
+
 const datasource = new DataSource({
     type: 'postgres',
     host: process.env.DB_HOST || 'localhost',
