+++ src/resources/album/entities/album.entity.ts
