+++ src/resources/artist/artist.controller.ts
