+++ src/resources/artist/artist.service.ts
