+++ src/resources/artist/dto/artist.dto.ts
