+++ src/resources/artist/dto/create-artist.dto.ts
