+++ src/resources/artist/dto/update-artist.dto.ts
