+++ src/resources/artist/entities/artist.entity.ts
