+++ src/resources/favorites/dto/add-favorite.dto.ts
