+++ src/resources/favorites/entities/favorite-response.entity.ts
