+++ src/resources/favorites/entities/favorite.entity.ts
