+++ src/resources/favorites/favorites.controller.ts
