+++ src/resources/favorites/favorites.module.ts
