+++ src/resources/favorites/favorites.service.ts
