+++ src/resources/user/entities/user.entity.ts
