+++ src/resources/user/user.module.ts
